requirements.txt
discord.py~=2.4
python-dotenv~=1.0
sqlalchemy[asyncio]~=2.0
aiosqlite~=0.20
